Restore `FuncBase.__instancecheck__` and fix decorators another way
@ilevkivskyi, I see your point about how overriding `__instancecheck__`
would create problems with debugging and `mypyc`, so I have implemented
an alternate solution. I have created a `get_callable` method to check
whether a node is callable and, in the case of a decorator, return an
instance of a separate `CallableDecorator` class that inherits from
`FuncItem`. I had originally thought to just have an `is_callable` method
that would test whether a `Node` was either an instance of `FuncBase`
or a callable `Decorator`, but this created problems with type resolution.
This solution is not exactly ideal either, but I think will create fewer
problems. I agree that the best solution would be to remove the `Decorator`
class altogether, but this would be a fairly major refactor. The problem
is that we do need a `Decorator` class in early passes before we have
type information, and we use the `Node.accept` method to add type information
to an instance, we don't have a simple way to replace a node with another
node of a different type. Let me know what you think of this approach.

One other stray observation:
* This fix requires that the decorator is explicitly annotated to return
  a `Callable`. It seems like this should be something we could determine
  through type inference, but I'm not sure how easy that would be to do.