Fix daemon support for our built-in plugins (#5828)
This is an initial take on support for proper support for the daemon from plugins
(or, at least from our existing built-in plugins).

There are two main support components for this:
* A new `plugin_generated` field on `SymbolTableNode` to indicate that a 
  node was created by a plugin and should be stripped as part of `aststrip`. 
  The `_add_method` support routine is modified to always set this field.
* An `add_plugin_dependency` method in `SemanticAnalyzerPluginInterface` 
   and a `plugin_deps` field in `MypyFile` for it to store dependencies.

The dataclass and attrs plugins are modified to use the `add_plugin_dependency` API.
Unfortunately, their use of it is somewhat subtle, requiring them to use wildcard deps.

The `plugin_generated` scheme, unfortunately, only kind of works, since plugins can
(and do) also *modify* nodes. The two approaches I can think of for fixing that are
making copies of ASTs and using that to ditch aststrip entirely (leads to a ~20% 
memory usage increase, but also lets us get rid of aststrip) or to have a callback
based system where plugins can register functions to rollback changes made by
the plugin (this is clunky and would be painful for plugin writers).
It isn't really clear that it matters, though, in practice. It might be workable to
(maybe with a little care) not need to strip those changes. It doesn't seem to
for `attrs` and `dataclasses` (which modify `Var`s to make them as properties
if the class is frozen.)

This doesn't add a dependency API to the typechecker plugins, since they
weren't needed. We might need to add one in the future.

Closes #5153.