Fix plugin hooks not being called correctly for dunder methods (#5700)
Fixes #4964.

* Fix method signature hooks not being called for dunder methods (#4964)

Previously, method signature hooks were only looked up and applied for
"real" method calls in the source code, but not for implicit dunder
method calls.

* Fix handling of plugin hooks for dunder methods (#4964)

Previously, most implicit dunder method calls were checked in two
separate steps: first the dunder method's signature was looked up on
the object's type, then the dunder method call was checked against that
signature. In the second step, the type of the self was not supplied
properly. Because of this, the method call was treated like a function
call by the plugin mechanism, and method hooks weren't looked up
properly.

This bug did not occur for binary operator dunder methods (such as
__add__); the binary operator checking code used helper methods to
supply the type of self correctly. These helper methods have been
generalized to work with any kind of method (not just binary operator
methods), and are now used consistently for all implicit method calls.

* Move method signature hook lookup/call into a helper method

Previously, method signatures were applied unconditionally in
check_call, however this caused them to be applied in places where they
shouldn't be (for example during overload resolution). This code has
been moved into a separate helper method, which is explicitly called in
the appropriate places.

[...technical commits...]

* Don't test that method hooks are called for __init__

To hook into class instantiation, a function hook for the class name
should be used instead, see discussion in #5700.