Add hooks for running tests on an installed version of mypy
The main motivation for all of this is to be able to validate
mypy_mypyc wheels.

Because the tests are an integrated subpackage of mypy, I couldn't
figure out a way to run the tests from the source directory against an
installed mypy, and I was reluctant to do a major overhaul of the
source organization, in part because some other projects take
advantage of some of the test infrastructure (mypyc, sqlalchemy-stubs).

So instead we run the tests from where they get installed with the
package. To support this, we allow overriding where the tests look for
test-data using a `MYPY_TEST_PREFIX` environment variable.