Fix some intermittent test failures due to failure to import plugin
While trying to test mypy_mypyc wheels in travis on OS X, I pretty
consistently got failures trying to import a plugin module. I had
trouble reproducing it anywhere else, but the issue turned out to be a
cousin of the dreaded stubgen flake (#4811), where caching in
importlib causes a file to be missed even though it is present now.

Here we solve it by using absolute paths in the `load_plugins`
manipulations of sys.path.