Add test case arg that skips path normalization
This commit lets us specify we want to skip path normalization
on some given test case, side-stepping the path-related errors
we were seeing.

Specifically, the problem was that mypy attempts to normalize
paths by replacing all instances of '\' with '/' so that the
output when running the tests on Windows matches the specified
errors. This is what we want to most of the time, except for a
few tests with Literals containing slashes.

I thought about maybe changing the output of mypy in general so
it always uses '/' for paths in error outputs, even on Windows:
this would mean we would no longer have to do path normalization.
However, I wasn't convinced this was the right thing to do: using
'\' on Windows technically *is* the right thing to do, and I didn't
want to complicate the codebase by forcing us to keep track of when
to use os.sep vs '/'.

I don't think I'll add too many of these test cases, so I decided
to just go with a localized solution instead of changing mypy's
error output.