Add basic end-to-end support for Literal types (#5947)
This pull request primarily adds support for string, int, bool, and
None Literal types in the parsing and semantic analysis layer. It also
adds a small smidge of logic to the subtyping and overlapping types
parts of the typechecking layer -- just enough so we can start using
Literal types end-to-end and write some tests. It also adds a 
"skip path normalization" option to our test cases.

Basically, my previous diff made a bunch of gross changes in a
horizontal way to lay out a foundation; this diff makes a bunch of
gross changes in a vertical way to create a proof-of-concept of the
proof-of-concept.

I'll probably need to submit a few follow-up PRs cleaning up some
stuff in the semantic analysis layer, but my hope is to switch focus
on fleshing out the type checking layer shortly after this PR lands.

Specific changes made:

1. I added a new 'RawLiteralType' synthetic type meant to represent any
   literal expression that appears in the earliest stages of semantic
   analysis. If these 'RawLiteralTypes' appear in a "Literal[...]" context,
   they transformed into regular 'LiteralTypes' during phase 2 of semantic
   analysis (turning UnboundTypes into actual types). If they appear
   outside of the correct context, we report an error instead.

   I also added a string field to UnboundType to keep track of similar
   information. Basically, if we have `Foo["bar"]`, we don't actually
   know immediately whether or not that "bar" is meant to be string
   literal vs a forward reference. (Foo could be a regular class or
   an alias for 'Literal').

   Note: I wanted to avoid having to introduce yet another type, so
   looked into perhaps attaching even more metadata to UnboundType
   or perhaps to LiteralType. Both of those approaches ended up looking
   pretty messy though, so I went with this.

2. As a consequence, some of the syntax checking logic had to move
   from the parsing layer to the semantic analysis layer, and some of
   the existing error messages changed slightly.

3. The PEP, at some point, provisionally rejected having Literals
   contain other Literal types. For example, something like this:

       RegularIds = Literal[1, 2, 3, 4]
       AdminIds = Literal[100, 101, 102]
       AllIds = Literal[RegularIds, AdminIds, 30, 32]

   This was mainly because I thought this would be challenging to
   implement, but it turned out to be easier then expected -- it
   happened almost by accident while I was implementing support
   for 'Literal[1, 2, 3]'.

   I can excise this part out if we think supporting this kind
   of thing is a fundamentally bad idea.

4. I also penciled in some minimal code to the subtyping and 
    overlapping types logic.

This diff also tweaks some of our test-case logic: we can now specify
we want to skip path normalization on certain test cases.

Specifically, the problem was that mypy attempts to normalize
paths by replacing all instances of '\' with '/' so that the
output when running the tests on Windows matches the specified
errors. This is what we want to most of the time: except when we want
to print out Literal types containing strings with slashes.

I thought about maybe changing the output of mypy in general so
it always uses '/' for paths in error outputs, even on Windows:
this would mean we would no longer have to do path normalization.
However, I wasn't convinced this was the right thing to do: using
'\' on Windows technically *is* the right thing to do, and I didn't
want to complicate the codebase by forcing us to keep track of when
to use os.sep vs '/'.

I don't think I'll add too many of these test cases, so I decided
to just go with a localized solution instead of changing mypy's
error output as a whole.