Implement a sqlite-based store for the cache
File system operations on OS X are pretty slow, and untarring a large
archive of mypy cache information can get pretty slow. Work around
this by using a sqlite database to store the entire cache in one file.

To do this we introduce a generic interface for storing metadata,
called `MetadataStore`. It presents an essentially key/value
interface. We provide two implementations, one using the existing file
system backing and one using sqlite.

It is enabled with the option `--sqlite-cache`, but is not the default
yet.

I'm not sure what the right thing to do about testing is. I've tested
it with the default changed, and everything passes.