Add tests for literals and generics
This pull request adds some tests verifying that basic interactions
between literals and generics work as expected.

It also mildly tweaks checkexpr so it actually considers the *whole*
type context instead of just part of it when deciding whether some
literal expression ought to be a Literal[...] type or not.