Add support for -posix and -windows test name suffixes
This allows writing platform-specific test cases, or
two platform-specific variants for a single test case.

A test case with -posix is only run on Linux and macOS, while
tests with a -windows suffix will only be run on Windows.

I want this for a test case I want to add where the output
depends on line ending style. There could be other uses
in the future.