Add tests for Literal types with incremental and fine-grained mode
This pull request adds a variety of tests for Literal types and
incremental and fine-grained mode.

Many of these tests are a bit perfunctory: literal types don't really
interact with incremental/fine-grained logic in a special way, so it was
a little hard for me to think of good test cases.

The only code-related change I ended up needing to make was to remove
the assert from the mypy.server.astmerge.TypeReplaceVisitor's
`visit_raw_literal_type` method.

This actually worries me a bit: we should have removed all instances of
RawLiteralType after phase 2 of semantic analysis. Does the merge step
happen before then? If the merge is supposed to happen after phase 2,
I can spend more time trying to track down the leak and update this PR.