Update tests to prepare for variable redefinition with new type
After the planned change, test cases no longer can rely on code
like this generating an error, since the two assignments create
two logically separate variables with independent types:

```
x = 0
x = ''  # Currently an error, but likely not in the future
```

Instead, we use an if statement to prevent redefinition.
This works since redefinition will only happen unconditionally:

```
x = 0
if int():
    x = ''  # Still an error
```

Even this may stop working in the future, and we may need to
another major test case update then. If we always use the
`if int():` idiom, finding things that need changing will be
easy.

In new test cases we should generally avoid using assignment
statements such as the above to generate errors. Instead, we
can use `reveal_type()` and calls to functions with incompatible
argument types, for example.