Switch to making final variables context-sensitive
This commit modifies this PR to make selecting the type of final
variables context-sensitive. Now, when we do:

    x: Final = 1

...the variable `x` is normally inferred to be of type `int`. However,
if that variable is used in a context which expects `Literal`, we infer
the literal type.

This commit also removes some of the hacks to mypy and the tests that
the first iteration added.