Add intelligent indexing of tuples, NamedTuples, and TypedDict (#6124)
This pull request adds a preliminary implementation of intelligent indexing of
tuples, NamedTuples, and TypedDicts. It uses the first approach we
discussed earlier: modifying the existing plugins and special-casing
code to also check if the expression has a Literal[...] type.

Once I'm finished with the baseline literal types implementation, I'll
look into circling back and seeing how viable the second approach is
(writing some sort of plugin that replaces the signatures of methods
like `.__getitem__` or `.get()` with overloads that use the appropriate
literal types).