Rename RawLiteralType to RawExpressionType
This commit renames "RawLiteralType" to "RawExpressionType". This is to
prepare for an upcoming commit, where we repurpose RawExpressionType to
represent any expression that does not convert cleanly into a type.

This lets us defer almost all error handling to the type analysis phase,
which makes it easier for us to generate cleaner error messages when
using Literal types.