Handle "invalid type" checks in semanatic analysis layer
This commit moves the "invalid type comments or annotations" error
handling out of the parsing stage into the type analysis segment of
the semantic analysis stage.

It does so by modifying the parser so it no longer reports an error
and returns an AnyType when encountering an invalid type: instead,
it returns a RawExpressionType. We then handle this RawExpressionType
in the semantic analysis layer in one of two ways:

- If the RawExpressionType corresponds to a literal and appears in a
  Literal[...] context, we generate a LiteralType.

- If the RawExpressionType does *not* correspond to a literal, or
  appears outside of a Literal[...] context, we report an error and
  produce an AnyType.

This commit also adjusts the "invalid type comments or annotations"
error message so it starts with an uppercase letter. This is more
consistent with the style of error messages the rest of the semantic
analysis layer typically returns.