Add docs for Literal types
This pull request adds some documentation on using Literal types.

It focuses mostly on how they can be parameterized, and different
ways of declaring variables to be Literal.

These docs deliberately do not mention the "intelligent indexing"
feature: we don't yet support intelligently indexing using variables
that are declared to be Final, so I think we should leave this
feature undocumented for now.