Add interactions between Literal and Final (#6081)
This pull request adds logic to handle interactions between Literal and
Final. In short, if the user were to define a variable like `x: Final = 3` and
latter do `some_func(x)`, mypy will attempt to type-check the code almost
as if the user had done `some_func(3)` instead.

This normally does not make a difference, except when type-checking code
using literal types. For example, if `some_func` accepts a `Literal[3]` up
above, the code would type-check since `x` cannot be anything other
then a `3`.

Or to put it another way, this pull request makes variables that use `Final`
with the type omitted context-sensitive.