Optionally allow redefinition of variable with different type (#6197)
If the flag `--allow-redefinition` is used, it's now possible to redefine a 
variable with a different type. For example, this code will be okay:

```
x = 0  # Define 'x' with type 'int'
print(x)
x = 'a'  # Define another 'x' with type 'str'
print(x)
```

The variable needs to be read before redefinition is possible. This is mainly 
to allow these idioms:

```
x: List[int]
x = []  # Not a redefinition

y = None  # type: List[int]  # Strict optional disabled
y = []  # Not a redefinition
```

It's also okay to use a type annotation in any of the assignments. This is
fine:

```
x: int = 0
print(x)
x: str = 'a'
print(x)
```

Redefinition is only allowed to happen in the same block and the nesting level
as the original definition. This lets us keep the implementation relatively simple,
and there will hopefully be less of a readability impact, since it's less likely that
users accidentally redefine a variable or use redefinition in confusing ways. We 
may want to lift the restriction in the future, however. Function arguments can
be redefined in the body but only in non-nested blocks.

The implementation does a simple static analysis to rename variables to make
them distinct when we detect redefinition. This happens before semantic 
analysis pass 1.

The new behavior is not fully backward compatible with the old behavior. For
example, here we lose type context in the second assignment:

```
x: List[int] = []
f(x)
x = []  # Need type annotation since this creates an independent variable
```

Internally we use a sequence of single quotes as a variable suffix
to represent renamed variants. The implementation provides an `unmangle()`
function to get the original name from a potentially mangled name.

It's still impossible to redefine final names, classes and certain other things.
The main use case is simple variables, either in single or multiple assignments.

I did some benchmarking (not with the latest version though) and the performance 
impact of the new pass was minor (under 2% IIRC).

Things to do in additional PRs:
* Add documentation.
* Make sure we unmangle names everywhere.
* Consider allowing redefinition to happen sometimes in another block. 
* Consider propagating type context from original definition.