New semantic analyzer: Support more than 2 passes
Currently additional passes are only performed to resolve
"from import" statements. They allow resolving indirect
(exported) imports.

Add a fixed maximum number of iterations that should always be
enough. If we hit the maximum, it's very likely that the remaining
unbound definitions are impossible to bind, or that there is a
definition cycle.