New semantic analyzer: Fix redefinitions
If the original definition could not be bound, disallow
further definitions, as the original should still take
precendence once it becomes ready.

The implementation uses a single set across potentially
nested scopes, which means that there can be unnecessary
aliasing between scopes. Since having the same name in
multiple scopes is pretty rare and since this only affects
performance, not correctness, I think that the simplified
implementation makes this a reasonable tradeoff.