New semantic analyzer: Support inheritance (#6249)
This adds support for incomplete references in base classes. If something
is incomplete, we don't create a TypeInfo yet and defer the current
target.

Most of the changes are related to calculating the relevant information
needed to decide if we need to defer without mutating the ClassDef
or having a TypeInfo.

Some things are not supported yet, such as defining class-based named
tuples or TypedDicts, and metaclasses. Also test coverage is minimal but
I think that this okay since we should be able to switch to using various
existing tests pretty soon.