New semantic analyzer: Optimize processing order of top levels
Previously we would process top levels in alternating orders if they
are deferred (a->b, b->a). This would result in the same module (b)
processed twice in a row, which may not be efficient if the deferral
is due to another module in the cycle, as the second step may then
make no progress.