New semantic analyzer: Support more features etc. (#6268)
This adds support for several new features:

 * Nested classes
 * Built-in type aliases such as List[T]
 * Overloaded functions
 * Decorated functions

This also includes some smaller fixes related to generic classes, adds
tests for other features, and makes the order of processing module top
levels consistent across iterations.

Previously we would process top levels in alternating orders if they
are deferred (a->b, b->a). This would result in the same module (b)
processed twice in a row, which may not be efficient if the deferral
is due to another module in the cycle, as the second step may then
make no progress.

Also remove fake definitions of these aliases from stubs, as
they are not necessary.