Optimize find_module by enumerating contents of directories on lib_path
In practice, almost all modules can be routed to the correct entry in
the lib_path by looking at just the first component of the module
name.

We take advantage of this by enumerating the contents of all of the
directories on the lib_path and building a map of which entries in
the lib_path could contain each potential top-level module that appears.

The previous cache for find_lib_path_dirs could co-exist with this,
but I removed it for simplicity since it didn't help anymore in my
measurements.