Mark abstract and protocol classes as not-instantiatable explicitly
Only show error when not-instantiatable class is instantiated. This will
allow plugins to allow instantiation of certain classes, even if they
are abstract or specify a protocol.

The particular problem that called for this solution:
[mypy-zope](https://github.com/Shoobx/mypy-zope) plugin
tries to implement zope interfaces as "protocols". Zope has the adaptation
pattern that looks like instantiation of an interface:

```python
adapter = IMyInterface(context)
```

since `IMyInterface` is marked as "protocol", mypy unconditionally rejects
such syntax with the message

```
error: Cannot instantiate protocol class "IMyInterface"
```

 The proposed change would allow zope plugin to mark
`IMyInterface` as both "protocol" and "instantiatable", effectively
disabling the error message.