New semantic analyzer: Add placeholder types etc. (#6302)
This adds the concept of placeholder type (and placeholder `TypeInfo`),
which are types that don't have a complete definition yet. They can be
useful in certain self-referential definition such as this:

```py
class C(Sequence[C]): ...
```

In the above example, we create a placeholder `TypeInfo` for `C`
before analyzing base classes. This allow `C` in `Sequence[C]` to
be bound. Each time there's a reference to a placeholder we defer
the current target so that a proper reference can be created on a
subsequent iteration. Placeholders should not leak semantic 
analysis and will generate an error during type checking.

They aren't that useful yet, but the goal is to eventually expand them
to support general recursive types.

This PR also includes a few additional small updates, including some
refactoring of how protocol types are detected (the refactoring isn't
directly related to the rest of the PR).