Fix dmypy on windows when compiled with mypyc
dmypy currently crashes on windows when compiled with mypyc
(since 0.660; it has *never* worked, I believe).

The root cause here is the abusive use of the platform variable,
which we set to 'mypyc', to tell bogus_type.py that mypyc is being used.
This (obviously) interferes with our platform checks, causing mypyc to
spuriously think code is unreachable, but because of the minutae of
how our platform checks were structured and which files are compiled,
this only breaks things for the windows daemon.

Switch to using always_true/always_false instead of abusing platform.