New semantic analyzer: check generic type arguments (#6395)
This adds a new pass that checks things like `int` being valid in
`C[int]`. We use a new pass since base classes can be incomplete
during semantic analysis, which breaks subtype checks that we
rely on here.

The implementation abstracts AST traversal into separate classes,
so that the implementation of the pass is relatively simple.

Fixes #6288.