New semantic analyzer: support assignment based TypedDict (#6439)
Fixes #6286

There are no tricks, it mostly matches one-to-one the implementation of named tuples. There is however one temporary hack, that we will be able to remove when we merge analyzers. Currently type checker depends on l.h.s pointing to a (redundant) `Var` that is always created by the old analyzer.