New semantic analyzer: Enable namedtuple tests and fix self-check error (#6452)
This fixes the `Name 'cls' already defined error` in self-check with new analyzer. The error is caused by the fact that now we actually analyze generated methods. I solve them simply by using underscored versions of `_self` and `_cls` (this is btw the same what runtime implementation does for some generated methods).

I also enable both named tuple test files. This needs the same temporary redundant variable hack as I did for typed dicts, it can be removed when we will not need to be compatible with old analyzer, I will create an issue so we will not forget about this.

I skip a bunch of test cases for recursive named tuples (there is already an issue for this). I also skip one test because it is tricky to fix, I will open a follow-up issue.