New semantic analyzer: enable tests and fix to fullnames (#6484)
This enables various tests for the new semantic analyzer. It also
fixes an issue with a local definition getting a fully qualified name.

Some tests have different output on new vs old semantic analyzer.
If the output on the new semantic analyzer was reasonable, I forced
the test to always run on it. In a few cases an existing issue had
been fixed on the new semantic analyzer.