Add info about classes and types to 'getting started' docs
While helping somebody use mypy, it came to my attention that our docs
never actually quite explain how classes and types interact.

This PR adds a new section named "Types and classes" to the "Getting
started" docs that tries to fix this.

This section focuses on introducing two ideas:

1. That you can use any Python class as a type
2. That if a variable has type T, it can also accept any instances of
   *subclasses* of type T.

The examples also give us an opportunity to sneak in a brief mini-lesson
on how to add type hints to classes, and what sorts of things do and
do not get inferred.

I also rewrote the next section about stubs/typeshed partly so it would
dovetail better with this new section but also partly because I always
felt that section was a little too dense and rushed.

Three other minor changes made:

- I added an example about sqlalchemy and sqlalchemy-stubs to one of the
  other pages.

- I removed a line directing people to the mypy issue tracker if they
  have questions.

- I dropped the term "library stubs" in favor of just "stub files".