Eliminate implicit reexports
This makes mypy check clean with the proposed --no-implicit-reexports
flag from PR #6562.

Some of the implicit reexports that were depended on were clearly
intended (type_visitor) and so made explicit, while some were clearly
unintended (Optional from mypy.types!) and fixed.

Probably worth landing this whichever way we decide on #6562, but I
think this PR is an argument in favor.