New semantic analyzer: fix type promotions special cases (#6609)
Set up type promotions after the main semantic analyzer pass to
avoid the need to defer. This fixes an issue with test stubs,
but the existing approach probably worked with typeshed.
Another benefit of this approach is that things are easier to
reason about, as the main semantic analysis pass is a bit
simpler.

This fixes the testDivmod type checker test case.