Allow ellipsis default args in "trivial" non-stub functions
Resolves https://github.com/python/mypy/issues/6531

This diff modifies the checker to allow default arguments to
be used in function definitions in non-stub files as long as the
body of that function is either empty or contains just a
`raise NotImplementedError`. (This is after ignoring docstrings.)

For rationale on why this is useful, see the issue linked above.

One note on function bodies containing only a `raise Blah`. I was
worried that letting mypy ignore the ellipsis when the function
raises *any* Exception would lead to unsafe behavior -- I didn't
want mypy to start allowing code that looks like this:

    def halt(self, reason: str = ...) -> NoReturn:
        raise MyCustomError("Fatal error: " + reason, self.line, self.current_context)

Whitelisting only NotImplementedError seemed like a good enough
compromise -- it's unlikely users will use the arguments in non-trivial
ways when raising that particular exception. It should also be
easy enough to relax this restriction in the future if necessary.