New semantic analyzer: fix redefining symbols as overloaded functions or classes (#6681)
Fixes https://github.com/python/mypy/issues/6432

This is a hot fix that simply extends the solution in PR https://github.com/python/mypy/pull/6429 to more situations. We agreed with @JukkaL that he will clean this up when he is back from PTO.

Note that we still give the errors for redefinitions that we consider invalid (see tests), but we don't crash now.