Make `narrow_declared_type` correctly handle Literal types
This diff fixes the `narrowed_declared_type` function to correctly
handle cases where the declared type is actually a `Literal[...]`.

This fixes https://github.com/python/mypy/issues/6952. In short, it
seems like using the new semantic analyzer somehow made mypy call
[the `narrow_type_from_binder(...)` function][0] where it previously
did not.

Specifically, mypy passed in the `var` expression and a `known_type` of
`Literal[42]`, we got a restriction of `builtins.int` (with a
`last_known_value` of 42), and attempted to narrow the `Literal[42]` on
line 3512.

But due to the missing case in `narrow_declared_type`, we ended up
returning just the int.

  [0]: https://github.com/python/mypy/blob/master/mypy/checkexpr.py#L3504