Keep track of fstring conversions and format-specs (#6965)
Modify fastparse.py to keep track of format specifiers and conversions for f-strings. 
Currently, braces values within f-strings are just converted to calls to format,
ignoring format specifiers and conversions ('{}'.format(...)). 
With this the information regarding format specifiers and conversions are passed along to format.
To accomodate format specifiers with expressions evaluated at runtime, we pass the
format specifier as an argument also (ex: '{!r{}}.format(..., "<30")').
Note that format specifiers can contain expressions that are evaluated at runtime.

Keeping track of this information is needed in order to provide mypyc with enough information to support fstrings with format specifiers and conversions.