Support flexible option specification in config files (#7054)
Support inverting options based on `no_`, and adding or removing a
`dis` prefix in the config file (and therefore in inline configuration also).

This moves the handling of `no_` prefixes from inline configuration to
the config file. As a result, the config file is brought closer into alignment
with the command line options.