Fix column numbers of type annotations
Previously many types didn't have column numbers.

This gives correct column number of simple types like
`Foo` and `List[Foo]` in Python 3 type annotations.

For type comments we cheat a bit and point them to the
surrounding statement or definition, which is still a
small improvement.

For some types we may still be missing column numbers,
but this covers the most common cases.