Point error message to incompatible argument expression
Previously we just pointed to the beginning of the call expression,
which is not great if there are multiple arguments.

Also point to incompatible list/dict/set item instead of the beginning
of a list/dict/set expression.

Note that this may require some `# type: ignore` comments to be
updated.