Remove 'is None' checks for MypyFile.path
I'm embarking on a (probably futile) quest to get mypy to type-check
cleanly under the new `--warn-unreachable` flag.

This pull request starts off by tackling some of the low-hanging fruit:
we have some unnecessary 'is None' checks for the MypyFile.path field,
which is annotated to be of just type 'str', not 'Optional[str]'.

As a side-effect, this diff fixes one itty-bitty bug: previously, we
would not generate reports when running something like
`mypy --any-exprs-report report_dir -c 'print("foobar")`.