modulefinder: use only dirs containing no __init__.py as namespace package
When used with --namespace-packages, a directory containing no __init__.pyi
was previously added as a candidate to near_misses even if it contained
a __init__.py. This logic doesn't seem sensible, particularly as this
directory was then chosen as the result of find_module() instead of
the __init__.py inside the directory.

This should cure #5854.