Support non-extension class compilation (mypyc/mypyc#657)
This PR introduces support for compiling classes as non-extension classes via calls to metaclass constructors. Using this non-extension mechanism, this PR supports compiling dataclasses (mypyc/mypyc#495) and Enums (and in theory, compilation of other classes that use metaclasses other than type). Currently, classes are compiled via this non-extension class mechanism if they are decorated, or have a user-specified metaclass other than ABCMeta, TypingMeta, or GenericMeta.

The current goal is to be able to use non-extension classes to support a de-optimized version of dataclasses and enums, and in the future, do some optimization work to make them faster.

Currently, non-extension classes do not support property setters and overloaded functions.