Report precision stats for imports and add new precision report (#7254)
Previously when reporting per-line type checking precision, imports were
treated as empty lines, which seems inconsistent. Now treat an import as
precise if the target module exists, and as imprecise if the target module is
missing from the build.

Also add a new report, `lineprecision`, which is mostly designed to be used
in tests. The output is more compact and cleaner than the existing
XML-based reports.

Make it possible to run report generation tests without using full stubs,
since using full stubs is slow.

This is still not quite perfect. If using module-level `__getattr__`, the import
should probably be treated as imprecise. Also, if a missing name is ignored
using # type: ignore, we may want to treat that as imprecise. For multi-line
imports, we only report precision for the first line. These would be a bit tricky
to fix in the report generator so I decided to skip these for now.

[Also fixes a small bug with % interpolation that I encountered.]