Config option to ignore file parse errors
**RFC quality. Not anywhere close to landing.**

The Mercurial project has some .py files that are Python 2 or
Python 3 only. Mypy's current behavior is to find all .py files
and attempt to parse them. If the parse fails, this is a blocking
error and execution aborts immediately. This behavior is obviously
not desirable for projects containing mixed Python version sources.

This commit teaches Mypy to make parse errors non-fatal/blocking.

It teaches the option system a new `ignore_parse_errors` boolean
to control whether parse errors are ignored. This option is
consulted at the low-level parser to determine whether to make
the parse error a blocker.

Open issues:

* Ignoring parse errors is a special case of "ignore file." Perhaps
  the more general feature would be better?
* The option doesn't seem to work on per-module entries in the
  mypy.ini file. I'm not sure why.
* Missing tests.
* Missing documentation updates.
* Is inserting an empty ast tree on parse failure the proper failure
  mode? I'm not sure what the implications are for using an empty
  tree versus ignoring the module completely.