Don't reexport star-imported symbols when implicit reexports are disabled
Currently, in regular files, symbols imported into module A with `from B
import *` are considered explicitly exported from module B as well, i.e.
they are considered part of module's B public API.

This behavior makes sense in stub files as a shorthand (and is specified
in PEP 484), but for regular files I think it is better to be explicit:
add the symbols to `__all__`.

Further discussion:
https://github.com/python/mypy/issues/7042#issuecomment-509005322