Fix spurious errors in inherited dataclasses in incremental mode
The dataclasses plugin incorrectly expected json to preserve
the ordering of its dictionaries, which led to spurious errors
about "Attributes without a default cannot follow attributes with one".

Worse, the errors are reported in the wrong file: they use the line
number of the attribute but the file of the dataclass being checked!

Fix the spurious error by serializing attributes in a list, since we
care about the order.

Reporting this error for attributes in parent classes is actually
useful, since multiple inheritance can cause this error. Report the
error at the class definition site causing the problem instead.
(The error message here could certainly be improved, but right
now I just want to fix the "wrong file" bugs, which are I think
the worst kind of bugs after crashes.)