Generalize the 'open' plugin for 'pathlib.Path.open' (#7643)
This pull request adds a plugin to infer a more precise
return type for the `pathlib.Path.open(...)` method.

This method is actually nearly identical to the builtin
`open(...)` method, with the only difference being that
`pathlib.Path.open(...)` method doesn't have the `file`
parameter. So, I refactored the logic in both plugins
into a shared helper method.