Fix operator error messages for __div__ and __cmp__ on python 2
Currently the code for turning method call errors into operator error
messages doesn't detect __div__ and __cmp__, so fix that.

One of my main motivations here is that this fixes type ignoring these
with an error code, since previously the real error would be
`arg-type` but the note would be `operator`.