Support FIXME names for unknown types in dmypy suggest (#7679)
This PR adds an option to `dmypy suggest` to replace `Any` in suggested signatures with a dummy name like `FIXME` or `UNKNOWN`. The main question here is whether we should replace only the top-level `Any`s , or also nested one. I am leaning towards the former, but this is not a strong opinion.