Avoid creating partial annotations (#7641)
This is important because `--disallow-incomplete-defs` is basically unusable otherwise, and I want to use it for e.g. HypothesisWorks/hypothesis#2071.

I'm well aware that this is a terrible hack, but it fixes #5954 for now.  @euresti had sketched out the proper fix but I just got lost in the codebase - hopefully someone will come back and fix it properly later, but for now this is good enough for me.