[mypyc] Support generating extension classes for dataclasses (#7817)
The approach here is to generate an extension class while also
populating a dictionary with the appropriate fields for the dataclass
constructor. Once the extension class is created, we swap out the
contents of its `__dict__` with the `__dict__` that `dataclass`
expects, invoke `dataclass`, and then put back all the original stuff.

Fixes #671.