mypyc: Fix C string encoding
Using repr() to encode C strings relied on implementation-defined
behaviour and had a number of bugs, for example:

* The C string could inadvertently contain trigraphs.

* Valid hexdecimal characters following a hexadecimal escape sequence
  would be parsed as part of the escape sequence.

Octal escape sequences are used for unprintable characters because they
do not have the same issue as hexadecimal escape sequences.

It would be possible to mitigate the issue with hexadecimal escape
sequences by emitting multiple string literals. However, the complexity
is not worth it, especially as the output is not designed to be
human-readable.