Fix 0.750 regression: union (non-)simplification should not affect inference (#8077)
Fixes https://github.com/python/mypy/issues/8051

The fix is as discussed in the issue. I didn't find a test case for the union `actual`, but I think it makes sense to simplify it as well. Note: I add the `pythoneval` test just as a regression test, if you think it is not worth it, I can remove it.