Don't mark a module as stale if its children change (#8134)
This was introduced in #1865 to fix an issue with new children modules
being added. I think we have fixed this issue in another way also
since then, since the important part of the test passes with the check
removed. (I added some more stuff to the test to increase my
confidence in this change).

I want to remove this check because it causes a *ton* of trouble with
our internal bazel integrations, where we try to generate cache
artifacts for each bazel target separately. This check means that
bazel targets for packages have their caches invalidated all the time,
since they were generated without the children.